Fixed tests and splitted into different files
